+++ src/Cake.Common/Text/TextTransformationAliases.cs
@@ -20,17 +20,65 @@ namespace Cake.Common.Text
         /// <param name="context">The context.</param>
         /// <param name="template">The template.</param>
         /// <returns>A <see cref="TextTransformation{TTemplate}"/> representing the provided template.</returns>
+        /// <example>  
+        /// This sample shows how to create a <see cref="TextTransformation{TTemplate}"/> using
+        /// the specified template.
+        /// <code> 
+        /// string text = TransformText("Hello &lt;%subject%&gt;!")
+        ///    .WithToken("subject", "world")
+        ///    .ToString();
+        /// </code> 
+        /// </example> 
         [CakeMethodAlias]
         public static TextTransformation<TextTransformationTemplate> TransformText(this ICakeContext context, string template)
+        {
+            return TransformText(context, template, "<%", "%>");
+        }
+
+        /// <summary>
+        /// Creates a text transformation from the provided template, using the specified placeholder.
+        /// </summary>
+        /// <param name="context">The context.</param>
+        /// <param name="template">The template.</param>
+        /// <param name="leftPlaceholder">The left placeholder.</param>
+        /// <param name="rightPlaceholder">The right placeholder.</param>
+        /// <returns>A <see cref="TextTransformation{TTemplate}"/> representing the provided template.</returns>
+        /// <example>  
+        /// This sample shows how to create a <see cref="TextTransformation{TTemplate}"/> using
+        /// the specified template and placeholder.
+        /// <code> 
+        /// string text = TransformText("Hello {subject}!", "{", "}")
+        ///    .WithToken("subject", "world")
+        ///    .ToString();
+        /// </code> 
+        /// </example> 
+        [CakeMethodAlias]
+        public static TextTransformation<TextTransformationTemplate> TransformText(
+            this ICakeContext context,
+            string template,
+            string leftPlaceholder,
+            string rightPlaceholder)
         {
             if (context == null)
             {
                 throw new ArgumentNullException("context");
             }
+            if (leftPlaceholder == null)
+            {
+                throw new ArgumentNullException("leftPlaceholder");
+            }
+            if (rightPlaceholder == null)
+            {
+                throw new ArgumentNullException("rightPlaceholder");
+            }
 
+            // Create the placeholder.
+            var placeholder = new Tuple<string, string>(leftPlaceholder, rightPlaceholder);
+
+            // Create and return the text transformation.
             return new TextTransformation<TextTransformationTemplate>(
                 context.FileSystem, context.Environment,
-                new TextTransformationTemplate(template));
+                new TextTransformationTemplate(template, placeholder));
         }
 
         /// <summary>
@@ -39,9 +87,46 @@ namespace Cake.Common.Text
         /// <param name="context">The context.</param>
         /// <param name="path">The template file path.</param>
         /// <returns>A <see cref="TextTransformation{TTemplate}"/> representing the provided template.</returns>
+        /// <example>  
+        /// This sample shows how to create a <see cref="TextTransformation{TTemplate}"/> using
+        /// the specified template file with the placeholder format <c>&lt;%key%&gt;</c>.
+        /// <code> 
+        /// string text = TransformTextFile("./template.txt")
+        ///    .WithToken("subject", "world")
+        ///    .ToString();
+        /// </code> 
+        /// </example> 
         [CakeMethodAlias]
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2202:Do not dispose objects multiple times", Justification = "Stream reader leaves stream open.")]
         public static TextTransformation<TextTransformationTemplate> TransformTextFile(this ICakeContext context, FilePath path)
+        {
+            return TransformTextFile(context, path, "<%", "%>");
+        }
+
+        /// <summary>
+        /// Creates a text transformation from the provided template on disc, using the specified placeholder.
+        /// </summary>
+        /// <param name="context">The context.</param>
+        /// <param name="path">The template file path.</param>
+        /// <param name="leftPlaceholder">The left placeholder.</param>
+        /// <param name="rightPlaceholder">The right placeholder.</param>
+        /// <returns>A <see cref="TextTransformation{TTemplate}"/> representing the provided template.</returns>
+        /// <example>  
+        /// This sample shows how to create a <see cref="TextTransformation{TTemplate}"/> using
+        /// the specified template file and placeholder.
+        /// <code> 
+        /// string text = TransformTextFile("./template.txt", "{", "}")
+        ///    .WithToken("subject", "world")
+        ///    .ToString();
+        /// </code> 
+        /// </example> 
+        [CakeMethodAlias]
+        [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2202:Do not dispose objects multiple times", Justification = "Stream reader leaves stream open.")]
+        public static TextTransformation<TextTransformationTemplate> TransformTextFile(
+            this ICakeContext context,
+            FilePath path,
+            string leftPlaceholder,
+            string rightPlaceholder)
         {
             if (context == null)
             {
@@ -51,18 +136,30 @@ namespace Cake.Common.Text
             {
                 throw new ArgumentNullException("path");
             }
+            if (leftPlaceholder == null)
+            {
+                throw new ArgumentNullException("leftPlaceholder");
+            }
+            if (rightPlaceholder == null)
+            {
+                throw new ArgumentNullException("rightPlaceholder");
+            }
 
             // Make the path absolute if necessary.
             path = path.IsRelative ? path.MakeAbsolute(context.Environment) : path;
+
+            // Create the placeholder.
+            var placeholder = new Tuple<string, string>(leftPlaceholder, rightPlaceholder);
 
             // Read the content of the file.
             var file = context.FileSystem.GetFile(path);
             using (var stream = file.OpenRead())
             using (var reader = new StreamReader(stream, Encoding.UTF8, true, 1024, true))
             {
+                // Create and return the text transformation.
                 return new TextTransformation<TextTransformationTemplate>(
                     context.FileSystem, context.Environment,
-                    new TextTransformationTemplate(reader.ReadToEnd()));
+                    new TextTransformationTemplate(reader.ReadToEnd(), placeholder));
             }
         }
     }