+++ src/Cake.Core/IO/SpecialPath.cs
@@ -6,6 +6,29 @@
     public enum SpecialPath
     {
         /// <summary>
+        /// The directory that serves as a common repository for application-specific 
+        /// data for the current roaming user.
+        /// </summary>
+        ApplicationData,
+
+        /// <summary>
+        /// The directory that serves as a common repository for application-specific 
+        /// data that is used by all users.
+        /// </summary>
+        CommonApplicationData,
+
+        /// <summary>
+        /// The directory that serves as a common repository for application-specific 
+        /// data that is used by the current, non-roaming user.
+        /// </summary>
+        LocalApplicationData,
+
+        /// <summary>
+        /// The Program Files folder.
+        /// </summary>
+        ProgramFiles,
+
+        /// <summary>
         /// The Program Files (X86) folder.
         /// </summary>
         ProgramFilesX86,
@@ -13,6 +36,6 @@
         /// <summary>
         /// The Windows folder.
         /// </summary>
-        Windows,        
+        Windows,
     }
 }