+++ ReleaseNotes.md
@@ -1,7 +1,15 @@
+### New in 0.1.34 (Released 2015/03/03)
+* Added support for NuGet SetApiKey.
+* Fixed unsafe logging.
+* Made text transformation placeholders configurable.
+* Added missing common special paths.
+* Fixed script path bug.
+* Added XML transformation support.
+
 ### New in 0.1.33 (Released 2015/02/24)
 * Added Multiple Assembly Support.
 * Added process output and timeout.
-* Fixed code generatiom issue.
+* Fixed code generation issue.
 * Added aliases for executing cake scripts out of process.
 * Added file hash calculator.
 * Added aliases for checking existence of directories and files.